Added generic GCN notice parser and fixed up tests for that
