+++ hermes/parsers/__init__.py
@@ -1,5 +1,4 @@
 from hermes.parsers.gcn_circular_parser import GCNCircularParser
-from hermes.parsers.gcn_lvc_counterpart_notice_parser import GCNLVCCounterpartNoticeParser
-from hermes.parsers.gcn_lvc_notice_plaintext_parser import GCNLVCNoticeParser
 from hermes.parsers.igwn_alert_parser import IGWNAlertParser
 from hermes.parsers.hermes_message_parser import HermesMessageParser
+from hermes.parsers.gcn_notice_plaintext_parser import GCNNoticePlaintextParser
